refactor && change redemptions radspecs
