+++ app/src/lib/utils.js
@@ -1,108 +1,3 @@
-import { round } from './math-utils'
-
-export function formatDecimals(value, digits) {
-  try {
-    return value.toLocaleString('latn', {
-      style: 'decimal',
-      maximumFractionDigits: digits,
-    })
-  } catch (err) {
-    if (err.name === 'RangeError') {
-      // Fallback to Number.prototype.toString()
-      // if the language tag is not supported.
-      return value.toString()
-    }
-    throw err
-  }
-}
-
-/**
- * Get the whole and decimal parts from a number.
- * Trims leading and trailing zeroes.
- *
- * @param {string} num the number
- * @returns {Array<string>} array with the [<whole>, <decimal>] parts of the number
- */
-function splitDecimalNumber(num) {
-  const [whole = '', dec = ''] = num.split('.')
-  return [
-    whole.replace(/^0*/, ''), // trim leading zeroes
-    dec.replace(/0*$/, ''), // trim trailing zeroes
-  ]
-}
-
-/**
- * Format a decimal-based number back to a normal number
- *
- * @param {string} num the number
- * @param {number} decimals number of decimal places
- * @param {Object} [options] options object
- * @param {bool} [options.truncate=true] Should the number be truncated to its decimal base
- * @returns {string} formatted number
- */
-export function fromDecimals(num, decimals, { truncate = true } = {}) {
-  const [whole, dec] = splitDecimalNumber(num)
-  if (!whole && !dec) {
-    return '0'
-  }
-
-  const paddedWhole = whole.padStart(decimals + 1, '0')
-  const decimalIndex = paddedWhole.length - decimals
-  const wholeWithoutBase = paddedWhole.slice(0, decimalIndex)
-  const decWithoutBase = paddedWhole.slice(decimalIndex)
-
-  if (!truncate && dec) {
-    // We need to keep all the zeroes in this case
-    return `${wholeWithoutBase}.${decWithoutBase}${dec}`
-  }
-
-  // Trim any trailing zeroes from the new decimals
-  const decWithoutBaseTrimmed = decWithoutBase.replace(/0*$/, '')
-  if (decWithoutBaseTrimmed) {
-    return `${wholeWithoutBase}.${decWithoutBaseTrimmed}`
-  }
-
-  return wholeWithoutBase
-}
-
-/**
- * Format the number to be in a given decimal base
- *
- * @param {string} num the number
- * @param {number} decimals number of decimal places
- * @param {Object} [options] options object
- * @param {bool} [options.truncate=true] Should the number be truncated to its decimal base
- * @returns {string} formatted number
- */
-export function toDecimals(num, decimals, { truncate = true } = {}) {
-  const [whole, dec] = splitDecimalNumber(num)
-  if (!whole && !dec) {
-    return '0'
-  }
-
-  const wholeLengthWithBase = whole.length + decimals
-  const withoutDecimals = (whole + dec).padEnd(wholeLengthWithBase, '0')
-  const wholeWithBase = withoutDecimals.slice(0, wholeLengthWithBase)
-
-  if (!truncate && wholeWithBase.length < withoutDecimals.length) {
-    return `${wholeWithBase}.${withoutDecimals.slice(wholeLengthWithBase)}`
-  }
-  return wholeWithBase
-}
-
 export function capitalizeFirst(string) {
   return string.charAt(0).toUpperCase() + string.slice(1)
-}
-
-export function formatTokenAmount(
-  amount,
-  isIncoming,
-  decimals = 0,
-  displaySign = false,
-  { rounding = 2 } = {}
-) {
-  return (
-    (displaySign ? (isIncoming ? '+' : '-') : '') +
-    formatDecimals(round(amount / Math.pow(10, decimals), rounding), 18)
-  )
 }
